moved iframe height hook to separate file
