+++ scopes/preview/ui/component-preview/preview.tsx
@@ -1,8 +1,9 @@
-import React, { useRef, IframeHTMLAttributes, useEffect, useState } from 'react';
+import React, { IframeHTMLAttributes } from 'react';
 import { ComponentModel } from '@teambit/component';
 import { usePubSubIframe } from '@teambit/pubsub';
 
 import { toPreviewUrl } from './urls';
+import useIframeContentHeight from './use-iframe-content-height';
 
 // omitting 'referrerPolicy' because of an TS error during build. Re-include when needed
 export interface ComponentPreviewProps extends Omit<IframeHTMLAttributes<HTMLIFrameElement>, 'src' | 'referrerPolicy'> {
@@ -58,50 +59,3 @@ export function ComponentPreview({
 ComponentPreview.defaultProps = {
   hotReload: true,
 };
-
-type CallbackFn = () => void;
-
-function useInterval(callback: CallbackFn, interval: number) {
-  const savedCallback = useRef<CallbackFn>(() => callback);
-
-  useEffect(() => {
-    savedCallback.current = callback;
-  });
-
-  // Set up the interval.
-  useEffect(() => {
-    function tick() {
-      savedCallback.current();
-    }
-    if (interval !== null) {
-      const id = setInterval(tick, interval);
-      return () => clearInterval(id);
-    }
-    // eslint-disable-next-line
-    return () => clearInterval();
-  }, [interval]);
-}
-
-export default function useIframeContentHeight({
-  interval = 250,
-  skip,
-}: {
-  interval?: number;
-  skip?: boolean;
-}): [React.MutableRefObject<HTMLIFrameElement | null>, number | undefined] {
-  const iframeRef: React.MutableRefObject<HTMLIFrameElement | null> = useRef(null);
-  const [iframeHeight, setIframeHeight] = useState(0);
-  if (skip) return [iframeRef, undefined];
-  useInterval(() => {
-    try {
-      const iframe = iframeRef.current;
-      // eslint-disable-next-line
-      const newHeight = iframe!.contentWindow!.document.body.scrollHeight;
-      setIframeHeight(newHeight);
-    } catch (_) {
-      // eslint-disable-next-line
-    }
-  }, interval);
-
-  return [iframeRef, iframeHeight];
-}
